Split GPU profiling into its own opengl implementation
