+++ src/profiling/mod.rs
@@ -3,6 +3,9 @@ mod profiling_disabled;
 #[cfg(feature = "profiling")]
 mod profiling_enabled;
 
+#[cfg(all(feature = "profiling", not(platform = "windows")))]
+mod opengl;
+
 #[cfg(not(feature = "profiling"))]
 pub use profiling_disabled::*;
 #[cfg(feature = "profiling")]
